pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "GwonDohyeon",email = "125738341+GwonDohyeon@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "djangorestframework (>=3.16.0,<4.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "django (>=5.2,<6.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "drf-yasg (>=1.21.10,<2.0.0)",
    "geopy (>=2.4.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
